+++ lib/auth.ts
@@ -9,7 +9,7 @@ import { Resend } from "resend";
 const isDev = process.env.NODE_ENV !== "production";
 
 // 30-minute session, no sliding refresh
-const ADMIN_SESSION_SECONDS = 60 * 15;
+const ADMIN_SESSION_SECONDS = 60 * 30;
 
 export const authOptions: NextAuthOptions = {
   adapter: PrismaAdapter(prisma),
@@ -20,6 +20,7 @@ export const authOptions: NextAuthOptions = {
     maxAge: ADMIN_SESSION_SECONDS, // expire after 30 min
     updateAge: 0,                  // don't extend on activity
   },
+
   // (also set JWT maxAge for good measure)
   jwt: {
     maxAge: ADMIN_SESSION_SECONDS,
@@ -34,10 +35,15 @@ export const authOptions: NextAuthOptions = {
         password: { label: "Password", type: "password" },
       },
       async authorize(creds) {
-        if (!isDev) return null; // safety: block in prod
+        // Safety: block in prod
+        if (!isDev) return null;
 
-        const email = creds?.email?.toLowerCase().trim();
+        // Narrow types so we have plain strings (not string | undefined)
+        const rawEmail = creds?.email?.toLowerCase().trim();
         const password = creds?.password?.trim();
+        if (!rawEmail || !password) return null;
+
+        const email = rawEmail;
 
         if (email === process.env.ADMIN_EMAIL && password === process.env.ADMIN_PASSWORD) {
           const user = await prisma.user.upsert({
@@ -45,8 +51,11 @@ export const authOptions: NextAuthOptions = {
             update: { role: "admin" },
             create: { email, role: "admin", name: "Admin" },
           });
-          return { id: user.id, email: user.email, name: user.name, role: user.role };
+
+          // Return a minimal user object for NextAuth (can include custom fields)
+          return { id: user.id, email: user.email, name: user.name ?? "Admin", role: user.role } as any;
         }
+
         return null;
       },
     }),
@@ -72,28 +81,35 @@ export const authOptions: NextAuthOptions = {
 
   callbacks: {
     async jwt({ token, user }) {
+      // Persist id/role on first sign-in
       if (user) {
-        token.id = (user as any).id ?? token.id;
-        token.role = (user as any).role ?? token.role;
+        (token as any).id = (user as any).id ?? (token as any).id;
+        (token as any).role = (user as any).role ?? (token as any).role;
       }
+
+      // If no user object (subsequent requests), hydrate from DB when we have an email
       if (!user && token?.email) {
         const u = await prisma.user.findUnique({
           where: { email: token.email as string },
           select: { id: true, role: true },
         });
         if (u) {
-          token.id = u.id;
-          token.role = u.role;
+          (token as any).id = u.id;
+          (token as any).role = u.role;
         }
       }
+
       return token;
     },
+
     async session({ session, token }) {
-      if (token?.id) (session.user as any).id = token.id;
-      if (token?.role) (session.user as any).role = token.role;
+      if ((token as any)?.id) (session.user as any).id = (token as any).id;
+      if ((token as any)?.role) (session.user as any).role = (token as any).role;
       return session;
     },
+
     async signIn({ user }) {
+      // Only allow sign-in if we have an email
       return !!user?.email;
     },
   },
